feat: switch to listings only
