+++ site/src/gatsby-plugin-theme-ui/index.js
@@ -180,7 +180,7 @@ export default {
 
   button: {
     // specificity haaaaaaaack
-    '&&': { color: 'background' },
+    '&&': { color: 'heading' },
     bg: 'primary',
     borderRadius: 6,
     display: 'inline-block',
